+++ src/array/ops/mod.rs
@@ -20,6 +20,7 @@ mod search_sorted;
 mod sort;
 mod sum;
 mod take;
+mod utf8;
 
 pub trait DaftCompare<Rhs> {
     type Output;
